requirements.txt
# Data processing and analysis
pandas>=2.2.0,<3.0.0
numpy==1.26.2

# File handling and data formats
python-dotenv==1.0.0
openpyxl==3.1.2
pyarrow==14.0.1

# Web scraping and HTML processing
bs4==0.0.2
requests==2.31.0
html2text==2024.2.26
lxml==5.1.0

# LangChain and related dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.28
langchain-text-splitters>=0.0.2
langchain-chroma>=0.1.3
langchain-openai>=0.2.0
openai>=1.0.0
chromadb==0.4.18
tiktoken>=0.7.0  # Updated this to match langchain-openai requirements
pydantic>=2.10.4

# Development tools
black==23.12.1
pylint==3.0.3
pytest==7.4.3

# LangGraph:
langgraph>=0.0.10          # Add this - core LangGraph package
tavily-python>=0.1.3       # Add this if using Tavily search
langgraph-checkpoint-sqlite
